+++ Release Notes.md
@@ -32,6 +32,7 @@ I've also adjusted the indentation of multiline function signature end parenthes
 
 ### 🗑 Deprecations
 
+* `CircularProgressView` has been deprecated, since it just applies a `.circular` style to `ProgressView`.
 * `ConditionalView` has been deprecated, since conditional views shouldn't be encouraged.
 * `ForEachEnumerated` has been deprecated. Use `Collection+EnumeratedArray` instead.
 * `OptionalView` has been deprecated, since conditional views shouldn't be encouraged.